requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2
torch==2.0.1
torchvision==0.15.2
PyWavelets==1.4.1
PyYAML==6.0
scikit-learn==1.3.0
tqdm==4.65.0
tensorboard==2.13.0
Pillow==10.0.0
